backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
google-genai
reportlab
